requirements.txt
antlr4-python3-runtime==4.9.3
audioread==3.0.1
biopython==1.79
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Cython==0.29.37
dataclassy==1.0.1
decorator==5.1.1
fonttools==4.53.0
greenlet==3.0.3
hdbscan==0.8.36
hydra-core==1.3.2
idna==3.7
importlib_metadata==7.1.0
joblib==1.4.2
kalpy-kaldi==0.0.0
kiwisolver==1.4.5
kneed==0.8.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.42.0
markdown-it-py==3.0.0
matplotlib==3.8.4
mdurl==0.1.2
Montreal_Forced_Aligner==3.1.1
msgpack==1.0.8
munkres==1.1.4
numba==0.59.1
numpy==1.26.4
omegaconf==2.3.0
packaging==24.0
pgvector==0.2.5
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
pooch==1.8.1
praatio==6.0.0
psycopg2==2.9.9
pycparser==2.22
Pygments==2.18.0
pynini==2.1.6
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rich-click==1.8.2
scikit-learn==1.2.2
scipy==1.13.1
setuptools==70.0.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.30
TextGrid==1.6.1
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
wheel==0.43.0
win-inet-pton==1.1.0
zipp==3.17.0
